niwang/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
crispy-bootstrap5==2024.10
defusedxml==0.7.1
dj-database-url==2.3.0
Django==5.1.4
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-formtools==2.5.1
django-heroku==0.3.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework-xml==2.0.0
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
gunicorn==23.0.0
httpie==3.2.4
idna==3.10
inflection==0.5.1
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
Pygments==2.19.1
PyJWT==2.10.1
PySocks==1.7.1
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
whitenoise==6.8.2
